+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer.Test/Helpers/CodeFixVerifier.Helper.cs
