+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer.Test/Helpers/DiagnosticVerifier.Helper.cs
