+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer.Test/MindTouchMaterializedEnumerableAnalyzer.Test.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{8B713886-9740-453A-B438-7EC8F68889DA}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>MindTouchMaterializeCollectionsAnalyzer.Test</RootNamespace>
-    <AssemblyName>MindTouchMaterializeCollectionsAnalyzer.Test</AssemblyName>
+    <RootNamespace>MindTouchMaterializedEnumerableAnalyzer.Test</RootNamespace>
+    <AssemblyName>MindTouchMaterializedEnumerableAnalyzer.Test</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
@@ -117,9 +117,9 @@
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\MindTouchMaterializeCollectionsAnalyzer\MindTouchMaterializeCollectionsAnalyzer.csproj">
+    <ProjectReference Include="..\MindTouchMaterializedEnumerableAnalyzer\MindTouchMaterializedEnumerableAnalyzer.csproj">
       <Project>{cefdd049-e7de-4503-ac24-2615eb6b1148}</Project>
-      <Name>MindTouchMaterializeCollectionsAnalyzer</Name>
+      <Name>MindTouchMaterializedEnumerableAnalyzer</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
