+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer.Test/Properties/AssemblyInfo.cs
@@ -4,11 +4,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("MindTouchMaterializeCollectionsAnalyzer")]
+[assembly: AssemblyTitle("MindTouchMaterializedEnumerableAnalyzer")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("MindTouch")]
-[assembly: AssemblyProduct("MindTouchMaterializeCollectionsAnalyzer")]
+[assembly: AssemblyProduct("MindTouchMaterializedEnumerableAnalyzer.Test")]
 [assembly: AssemblyCopyright("Copyright ©  2016")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -27,5 +27,6 @@ using System.Runtime.InteropServices;
 //
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
-[assembly: AssemblyVersion("1.0.*")]
+// [assembly: AssemblyVersion("1.0.*")]
+[assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyFileVersion("1.0.0.0")]
