+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer.Test/Verifiers/CodeFixVerifier.cs
