+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{CEFDD049-E7DE-4503-AC24-2615EB6B1148}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>MindTouchMaterializeCollectionsAnalyzer</RootNamespace>
-    <AssemblyName>MindTouchMaterializeCollectionsAnalyzer</AssemblyName>
+    <RootNamespace>MindTouchMaterializedEnumerableAnalyzer</RootNamespace>
+    <AssemblyName>MindTouchMaterializedEnumerableAnalyzer</AssemblyName>
     <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <TargetFrameworkProfile>Profile7</TargetFrameworkProfile>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
