+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer/README.md
@@ -1,4 +1,4 @@
-﻿MindTouchMaterializeCollectionsAnalyzer
+﻿MindTouchMaterializedEnumerableAnalyzer
 =======================================
 
 We follow the convention that whenever an IEnumerable type is passed into a function or returned from a function, it must be materialized to a concrete type. For example:
