+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer/packages.config
