+++ MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer/tools/install.ps1
