+++ README.md
@@ -10,10 +10,10 @@ MindTouchEnumSwitchAnalyzer
 
 Roslyn analyzer that ensures that all enum cases are handled when used inside a switch statement. [Read more.](MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/README.md)
 
-MindTouchMaterializeCollectionsAnalyzer
+MindTouchMaterializedEnumerableAnalyzer
 ---------------------------------------
 
-Analyzer that ensures that any collections passed in as parameters to a function, or returned from a function are materialized. [Read more.](MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer/README.md)
+Analyzer that ensures that any collections passed in as parameters to a function, or returned from a function are materialized. [Read more.](MindTouchMaterializedEnumerableAnalyzer/MindTouchMaterializedEnumerableAnalyzer/README.md)
 
 System Requirements
 -------------------
